Fix UI build in docker container

+++ scripts/build-ui.sh
@@ -88,8 +88,14 @@ sed -i -e \
   "s/showFrameworkMode: false/showFrameworkMode: true/g" \
   $frontendFolder/src/config.js
 
+rm -f "$ROOT/.eslintrc"
+
 npm i --no-audit
 npm run build
+
+if ! [ -s "$frontendFolder/build/index.html" ]; then
+  exit 1
+fi
 
 mv -f $frontendFolder/build/* $uiBuildFolder
 cp $pathToTriggerElectronLoad $uiBuildFolder/triggerElectronLoad.js
